fix: ollama fix, import fixes, upload fix

+++ backend/embeddings.py
@@ -4,6 +4,6 @@ model = SentenceTransformer('all-MiniLM-L6-v2')
 
 def generate_embeddings(clauses):
 
-    embeddings = model.encode(clauses)
+    embeddings = model.encode(clauses).astype('float32')
 
     return embeddings
